LTD-587 Keep track of seen / processed emails in inbox
This commit changes the way the service processes emails. Instead of constantly
looking at the last 3, it creates an iterator of unread messages in the mailbox.
Once they are successfully dealt with they are marked as 'READ'. If they
couldn't be processed then they are marked as 'UNPROCESSABLE'. This will allow
us to continue processing messages in the mailbox if there are errors and also
allow us to mark 'UNPROCESSABLE' messages as 'UNREAD' once the reason for them
being unprocessable is fixed.