Convert the list of message_ids read to string
message_id is a UUIDfield in MailReadStatus model so when we filter the list of
messages that are READ/UNPROCESSABLE they come out as UUIDs but the message_id
extracted from the message header is also uuid in the form of a string.

'f47e773a-9bce-1f61-483b-368887650c96' not in {UUID('f47e773a-9bce-1f61-483b-368887650c96')}

Because of this the above condition becomes True so we endup creating another entry
in the MailReadStatus table and try to process the same mail again.