Don't block send lite updates to HMRC when there is a rejected mail
When sending LITE updates to HMRC we check if there is any pending email
or any email got rejected. If we find any rejected mail in the list of mails
sent then sending of updates is not performed effectively blocking all
licence updates from LITE. Also to resolve a rejection we actually need
to send another email which is not possible with this restriction hence
this is lifted completely.

It is to be noted that SPIRE updates are done in a separate task and they are
not blocked.