Check only the last 500 emails when looking for new emails in the mailboxes
Since we don't delete emails from these mailboxes the total number can be
very high over a period which increases the processing time.

The list we pull from pop3 gives us message number and size eg b"552 42569".
As we don't delete emails the message number is an increasing value so the
new emails will always be at the end and checking last emails should be
sufficient.