Fix sending of pending emails stuck in the queue
Usually we send one mail to HMRC and wait for the reply. Sometimes we get
multiple emails (eg usageData mail along with a licence data or reply)
in the same polling session then we process both of them (serialize and save)
but currently we send only one of them. The status of second one becomes "pending".

If there are no new emails currently we simply return and the pending mail
doesn't get processed. If the pending mail is of type licenceData then we get into
a deadlock situation because the mail is not sent to HMRC and we don't get any
reply. Unless we get a reply (new mail) we don't process these emails (currently).

To fix this we check for any pending mails when there are no unread mails in the inbox.
Please note that even if there are multiple pending mails we only process one of them.
If there are only pending usageData mails we can send all of them at once but there
can be a combination of usageData and licenceData as explained below so it is safe
to send one of them and the remaining gets cleared in next polling session.

Possible scenarios are,

Case1: When only one licenceData mail is pending

| Mail id      | edi_filename                                    | status         |
+ -------------+-------------------------------------------------+----------------+
| ac1323b135fb | CHIEF_LIVE_SPIRE_licenceData_98720_202106180855 | reply_sent     |
| d2d9b8d1f582 | CHIEF_LIVE_SPIRE_licenceData_98721_202106180855 | reply_sent     |
| dc1323b1abcd | CHIEF_LIVE_SPIRE_licenceData_98722_202106180855 | pending        |

Case2: When only usageData mails are pending. Multiple mails are possible if none
are received over the weekend and both arrived on monday.

| Mail id      | edi_filename                                    | status         |
+ -------------+-------------------------------------------------+----------------+
| ac1323b135fb | CHIEF_LIVE_SPIRE_licenceData_98720_202106180855 | reply_sent     |
| d2d9b8d1f582 | CHIEF_LIVE_SPIRE_licenceData_98721_202106180855 | reply_sent     |
| dc1323b1abcd | CHIEF_LIVE_SPIRE_usageData_5050_202106180855    | pending        |
| bb2223b132bc | CHIEF_LIVE_SPIRE_usageData_5051_202106180855    | pending        |

Case3: When both licenceData and usageData mails are pending. This is possible if
we haven't received any usageData files over the weekend and on monday they have
arrived at the same time of a new licenceData file. This is unlikely but possible.

| Mail id      | edi_filename                                    | status         |
+ -------------+-------------------------------------------------+----------------+
| ac1323b135fb | CHIEF_LIVE_SPIRE_licenceData_98720_202106180855 | reply_sent     |
| d2d9b8d1f582 | CHIEF_LIVE_SPIRE_licenceData_98721_202106180855 | reply_sent     |
| dc1323b1abcd | CHIEF_LIVE_SPIRE_usageData_5050_202106180855    | pending        |
| bb2223b132bc | CHIEF_LIVE_SPIRE_usageData_5051_202106180855    | pending        |
| 541323b18976 | CHIEF_LIVE_SPIRE_licenceData_98722_202106180855 | pending        |

Another variation of this case is,

| Mail id      | edi_filename                                    | status         |
+ -------------+-------------------------------------------------+----------------+
| ac1323b135fb | CHIEF_LIVE_SPIRE_licenceData_98720_202106180855 | reply_sent     |
| d2d9b8d1f582 | CHIEF_LIVE_SPIRE_licenceData_98721_202106180855 | reply_sent     |
| dc1323b1abcd | CHIEF_LIVE_SPIRE_usageData_5050_202106180855    | reply_sent     |
| 541323b18976 | CHIEF_LIVE_SPIRE_licenceData_98722_202106180855 | pending        |