Remove unique_together constraint on TransactionMapping model
This model is used to record the usage data transactions for a given licence.
For reasons unknown there is a unique_together constraint enforced on this based
on certain fields even though we are not using this model for any decision
making etc.

If usage is recorded for any good of a licence then the usage data file coming
from hmrc is containing multiple licenceUsage lines in the file and because of
the unique_together constraint we were not able to separate LITE specific
usage data and record it correctly in our system.

Fixing this by removing the constraint, this should allow us to record the
usage for LITE licences correctly.

Please note this won't affect any of the SPIRE licences.