Don't send another licenceData mail if waiting for a reply
We should only send licenceData mail to HMRC after we receive the reply for the previous mail.
If we are waiting for a reply we shouldn't send another mail.

Currently because of this bug the service is sending mail even when we are waiting for reply.
We can however send usageData mails to SPIRE so they are not prevented.