ICMSLST-1285 Switch to using Django's EmailMessage class (#5)
Instead of using the MailServer class for both POP3 and SMTP, use Django's
mail backends for doing the SMTP stuff. This involves removing SMTP from
MailServer, adding a helper to get a Django mail backend configured with
different SMTP settings, and constructing a Django EmailMessage instance
instead of a plain Python message instance.

Then when sending email, we need to get the appropriate Django mail backend
and use that to send the Django EmailMessage.

When running tests Django creates a locmem SMTP backend which never sends
nothing, but we can access django.core.mail.outbox to see what is "sent".
Mailhog is not required to run the tests (but is still super useful for
local development).

To simplify setup, several project settings have been made optional, with
default values that allow the application to start, which will then need
setting when deployed to production.