ICMSLST-1269 Remove hard-coded path from the LITE_API_URL setting (#6)
When the icms-hmrc app calls back to the main icms app with the license
results, the URL path was hard-coded to /licenses/hmrc-integration/. But
what if we don't want that path? This change uses the LITE_API_URL setting
as-is for the call-back URL, so you can specify any path.