Add management command to resend an email corresponding to the given HMRC run number
One of the support task is to manually resend an email to either HMRC/SPIRE.
This involves logging into the mailbox, identify the correct mail using the
run number, downloading the attachment and resending it. This is not efficient
and error prone so an management command is added to pick up the correct
mail as per the given run number and resend it.

Depending on the run number, mail extract type and status the command determines
the destination and sends it again.

v2: Fix review comments
Use old style formatting for logging as it will defer formatting if its not required
Correctly store the dry_run command line argument instead of as string