ICMSLST-1429 Merge (#8)
* Add field to skip licenses blocking the queue.

* Add filter to skip failing licenses.

* Install dev dependencies for docker image

* Install ipdb in dev requirements

* Handle removal of extra line carriages in edi data

* Add default value for skip field

This should already be handled by the default value of this field in the model definition however this is raising errors with this field not being fully populated so adding this more explicit value to get around this

* Remove legacy tag for new test

* Add management command to resend an email corresponding to the given HMRC run number

One of the support task is to manually resend an email to either HMRC/SPIRE.
This involves logging into the mailbox, identify the correct mail using the
run number, downloading the attachment and resending it. This is not efficient
and error prone so an management command is added to pick up the correct
mail as per the given run number and resend it.

Depending on the run number, mail extract type and status the command determines
the destination and sends it again.

v2: Fix review comments
Use old style formatting for logging as it will defer formatting if its not required
Correctly store the dry_run command line argument instead of as string

* Downgrade psycopg2 due to compatibility issues

* Add logging to investigate dummy email issues

* Return anonymous user object from hawk auth

This was incorrectly returning the class whereas a user object is expected

* Update logging calls to add stack trace appropriately

* Add a logging error call for licence usage payload issues

* Update logging calls from stack_info to exc_info

This is due to the fact that sentry doesn't display the stack trace when just using stack_info but will always show a trace (even without an exception) if you set exc_info

* Add tests for checking background tasks begin

* Switch on tz in local example env

* Remove mailhog from background test runner

* degrade unknown emails from error to warning

* stripping territory code while building the data file so as to push licences that come with it

* add pre-commit files

* update pipenv

* ICMSLST-1429 Fix SMTP-related tests

The tests broke because we merged changes from upstream, which uses a single
MailServer class to handle POP3 and SMTP. But in icms-hmrc the SMTP stuff
has been moved to a Django mail back-end. Changing the new code to use
the default SMTP connection fixes the tests.

Co-authored-by: Ali Zaidi <alixedi@gmail.com>
Co-authored-by: Kevin Carrogan <kevin.carrogan@gmail.com>
Co-authored-by: Arun Siluvery <arun.siluvery@digital.trade.gov.uk>
Co-authored-by: Sekhar Panja <sekhar.panja@digital.trade.gov.uk>
Co-authored-by: depsiatwal <depsi.a@gmail.com>
Co-authored-by: Gurdeep Atwal <44236490+depsiatwal@users.noreply.github.com>