ICMSLST-1464 Fix broken use of logging (#99)
These logging calls had a mix of f-strings and format calls, which "works"
but does not do what was intended.

    >>> f"Status {0}".format("fail")
    'Status 0'

But for Python's logging module, one should use positional arguments, not
string formatting (f-strings or other flavours). This way the logging
module can delay string formatting until it is needed (or even never if
the logging level means the message is skipped). Also the logging module
will swallow unicode encode/decode errors when formatting strings, which
allows the program to continue.