Remove skip field in LicencePayload model
This field was added as a temporary fix to skip licence details that lite-hmrc
was failing to send where the country code length was more than the specified
limit. We fixed the underlying issue and also resent the skipped licence details
manually to HMRC and so this field is no longer required.

Note that this only affects licences issued from LITE.