Skip reporting usage data when there are no licences
We parse Usage data transactions and group them based on licence status.
We only process usage data of licences that are open and the data for
completed licences are ignored. So if there is usage data we expect
the licence count to be non-zero. If we find it as 0 which is very
unlikely in which case we log the error and skip processing it.