ICMSLST-1423 Refactor line number logic in CHIEF message builder (#107)
Every line in a CHIEF message starts with the line number, and "end" lines
also have a count of the number of lines since the opening transaction line
(named `distance` in this code change). Moving the logic for line numbering
to the end simplifies the code.

Longer term we should break the logic in `licences_to_edifact(..)` into
smaller pieces, and add flexibility for sending with ILBDOTI (the ICMS
system) and for import licence requests (not just `usage_code = "E"`).