Remove unnecessary mail is pending error as it is not an error
When send_licence_data_to_hmrc() task runs it checks to see if there is an update
in progress. An update can be either we are waiting for a reply or the next email
is about to be processed (eg new email just got created and check_inbox task about
to process it, in this case it will be in pending state).

Since we have to wait till we completely process an email (i.e, reply_sent state)
we should skip this update when there is an update in progress, this gets picked
up in the next polling cycle.

Currently we are logging an error if an update in progress. This is not an error
and it unnecessarily adds noise to the alerts hence remove this.