ICMSLST-1541 Add types for CHIEF message lines (#119)
This uses dataclasses to define the line structure, because they support
inheriting field declarations whereas collections.namedtuple does not.
Dataclasses also allow using keyword arguments with the constructor.

The big improvement is we can have a little bit more self-documenting
code with names for the fields, instead of items in tuples with an
order that seems arbitrary but is very significant.

In CHIEF syntax every line starts with the line number and record type,
but for our use it is very useful to leave the line number unknown
until we construct the final message. But if you have one field in a
dataclass as optional, then all fields that follow must also be optional.