ICMSLST-1537 Use gunicorn for serving in production
Instead of Django's development server, use gunicorn. Default behaviour
reads a configuration from gunicorn.conf.py and binds to 0.0.0.0:$PORT
when $PORT is defined in the environment.

The config overrides the `Server` response header to hide gunicorn's
signature. This is silly, but part of a belt-and-braces approach to
security. However the header is then reset by another proxy. If we ever
deploy the app with a proxy that respects the app `Server` header, it
would show the customised, obscured signature.