ICMSLST-1578 Build CHIEF messages for other systems (e.g. ICMS)
Now that ICMS is using the same code for sending licence requests to CHIEF, we
need to allow other system identifiers in the CHIEF request heading.

This isn't fully flexible in that it still assumes that all _forwarded_
messages are from SPIRE, but that isn't currently a problem because the ICMS
team is not intending to support forwarding messages for legacy ICMS, so won't
be running in parallel with the legacy system (and won't forward things).