ICMSLST-1634 Remove hard-coded path from the LITE_API_URL setting (again)
Allow a full path such as `LITE_API_URL=http://foo/bar/`, but fall back to the
current path value. So `LITE_API_URL=http://foo` results in a call to
`http://foo/licences/hmrc-integration/`.

Removing the hard-coded path means the ICMS team can deploy the same code but
choosing a different path on the server targeted for the licence callback.

This was previously done in ICMSLST-1269, but this time it is backwards
compatible so the existing settings work without changes.