ICMSLST-1637 Split SMTP settings from POP3 settings
There are multiple mail server settings supported so that we can check
different POP3 accounts, but that assumed the SMTP server was the same
as the POP3 server. In practice we always use the same SMTP server and
the same "From:" address when sending a message.

Change the settings so that SMTP things are independent of the POP3
things, remove the `MailServer.use_tls` attribute (it was unused).

The environment variables used for configuring SMTP are the same:

- EMAIL_HOSTNAME
- EMAIL_SMTP_PORT
- EMAIL_USE_TLS
- EMAIL_USER
- EMAIL_PASSWORD

These environment variables are now unused and can be removed:

- INCOMING_EMAIL_SMTP_PORT
- HMRC_TO_DIT_EMAIL_SMTP_PORT
- MOCK_HMRC_EMAIL_SMTP_PORT
- SPIRE_STANDIN_EMAIL_SMTP_PORT